Move TwoPointTheory to models.two_point
